service-workers/service-worker/partitioned-service-worker-getRegistrations.tentative.https.html

<!DOCTYPE html>
<meta charset="utf-8"/>
<title>Service Worker: Partitioned Service Workers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script src="/common/get-host-info.sub.js"></script>

<body>

<script>
  // Idea for this test:
  // Similar to the other one, create a cross-site window.
  // On both the window and the test you should getAllRegistrations and confirm
  // (Note: we have to waitUntil() on a promise or the SW might be killed)
  // 1) There is only 1 seen
  // 2) The SW ID for the site and cross-site versions are different.

  // The resolve function for the current pending event listener's promise.
// It is nulled once the promise is resolved.
var message_event_promise_resolve = null;

function messageEventHandler(evt) {
  if (message_event_promise_resolve) {
    local_resolve = message_event_promise_resolve;
    message_event_promise_resolve = null;
    local_resolve(evt.data);
  }
}

function makeMessagePromise() {
  if (message_event_promise_resolve != null) {
    // Do not create a new promise until the previous is settled.
    return;
  }

  return new Promise(resolve => {
    message_event_promise_resolve = resolve;
  });
}

function loadAndReturnSwData(t, url, frame_type) {
  if (frame_type !== 'iframe' && frame_type !== 'window') {
    return;
  }

  const message_promise = makeMessagePromise();

  // Create the iframe or window and then return the promise for data.
  if ( frame_type === 'iframe' ) {
    const frame = with_iframe(url, false);
    t.add_cleanup(async () => {
      const f = await frame;
      f.remove();
    });
  }
  else {
    // 'window' case.
    const w = window.open(url);
    t.add_cleanup(() => w.close());
  }

  return message_promise;
}

promise_test(async t => {
  const script = './resources/partitioned-storage-sw.js'
  const scope = './resources/partitioned-'

  // Add service worker to this 1P context.
  const reg = await service_worker_unregister_and_register(t, script, scope);
  t.add_cleanup(() => reg.unregister());
  await wait_for_state(t, reg.installing, 'activated');

  // Register the message listener.
  self.addEventListener('message', messageEventHandler);
  t.add_cleanup(() =>{
    self.removeEventListener('message', messageEventHandler, false);
  });

  // Open an iframe that will create a promise within the SW.
  // The query param is there to track which request the service worker is
  // handling.
  //
  // This promise is necessary to prevent the service worker from being
  // shutdown during the test which would cause a new ID to be generated
  // and thus invalidate the test.
  const wait_frame_url = new URL(
    './resources/partitioned-waitUntilResolved.fakehtml?From1pFrame',
    self.location);

  // We don't really need the data the SW sent us from this request
  // but we can use the ID to confirm the SW wasn't shut down during the
  // test.
  const wait_frame_1p_data = await loadAndReturnSwData(t, wait_frame_url,
                                                       'iframe');

  // Now we need to create a third-party iframe that will send us its SW's
  // ID.
  const third_party_url = new URL(
    './resources/partitioned-service-worker-third-party-window.html'
    + '?partitioned-service-worker-third-party-iframe-getRegistrations.html',
    get_host_info().HTTPS_NOTSAMESITE_ORIGIN + self.location.pathname);

  // Create the 3p window (which will in turn create the iframe with the SW)
  // and await on its data.
  const frame_3p_ID = await loadAndReturnSwData(t, third_party_url, 'window');

  // Now get this frame's SW's ID.
  const frame_1p_ID_promise = makeMessagePromise();

  const retrieved_registrations =
        await navigator.serviceWorker.getRegistrations();


  // Register a listener on the service worker container and then forward to
  // the self event listener so we can reuse the existing message promise
  // function.
  navigator.serviceWorker.addEventListener('message', evt => {
    self.postMessage(evt.data, '*');
  });

  retrieved_registrations[0].active.postMessage({type: "get-id"});

  const  frame_1p_ID = await frame_1p_ID_promise;

  // First check that the SW didn't shutdown during the run of the test.
  assert_true(wait_frame_1p_data.ID === frame_1p_ID.ID,
    "1p SW didn't shutdown");
  // Now check that the 1p and 3p IDs differ.
  assert_false(frame_1p_ID.ID === frame_3p_ID.ID,
    "1p SW ID matches 3p SW ID");

  // Finally, for clean up, resolve the SW's promise so it stops waiting.
  const resolve_frame_url = new URL(
    './resources/partitioned-resolve.fakehtml?From1pFrame', self.location);

  // We don't care about the data.
  await loadAndReturnSwData(t, resolve_frame_url, 'iframe');

}, "ServiceWorker's getRegistration() is partitioned");


</script>

</body>
